backend/requirements.txt
flask==2.0.2
flask-restful==0.3.9
flask-cors == 4.0.0
mysql-connector-python==8.0.28
Werkzeug==2.0.2
